move httpclient factory to web tools
