+++ compass-geocoder-web/src/commonMain/kotlin/dev/jordond/compass/geocoder/web/Geocoder.kt
@@ -38,8 +38,8 @@ public fun Geocoder(
 public fun Geocoder(
     forwardEndpoint: ForwardEndpoint,
     reverseEndpoint: ReverseEndpoint,
-    json: Json = HttpApiPlatformGeocoder.json(),
-    httpClient: HttpClient = HttpApiPlatformGeocoder.httpClient(json),
+    json: Json = HttpApiEndpoint.json(),
+    httpClient: HttpClient = HttpApiEndpoint.httpClient(json),
     dispatcher: CoroutineDispatcher = Dispatchers.Default,
 ): Geocoder {
     val apiGeocoder = HttpApiPlatformGeocoder(forwardEndpoint, reverseEndpoint, json, httpClient)
@@ -57,8 +57,8 @@ public fun Geocoder(
  */
 public fun ForwardGeocoder(
     endpoint: ForwardEndpoint,
-    json: Json = HttpApiPlatformGeocoder.json(),
-    httpClient: HttpClient = HttpApiPlatformGeocoder.httpClient(json),
+    json: Json = HttpApiEndpoint.json(),
+    httpClient: HttpClient = HttpApiEndpoint.httpClient(json),
     dispatcher: CoroutineDispatcher = Dispatchers.Default,
 ): ForwardGeocoder {
     return ForwardGeocoder(ForwardHttpApiPlatformGeocoder(endpoint, httpClient), dispatcher)
@@ -75,8 +75,8 @@ public fun ForwardGeocoder(
  */
 public fun ReverseGeocoder(
     endpoint: ReverseEndpoint,
-    json: Json = HttpApiPlatformGeocoder.json(),
-    httpClient: HttpClient = HttpApiPlatformGeocoder.httpClient(json),
+    json: Json = HttpApiEndpoint.json(),
+    httpClient: HttpClient = HttpApiEndpoint.httpClient(json),
     dispatcher: CoroutineDispatcher = Dispatchers.Default,
 ): ReverseGeocoder {
     return ReverseGeocoder(ReverseHttpApiPlatformGeocoder(endpoint, httpClient), dispatcher)
