+++ compass-geocoder-web/src/commonMain/kotlin/dev/jordond/compass/geocoder/web/HttpApiPlatformGeocoder.kt
@@ -5,15 +5,7 @@ import dev.jordond.compass.Place
 import dev.jordond.compass.geocoder.PlatformGeocoder
 import dev.jordond.compass.tools.web.HttpApiEndpoint
 import io.ktor.client.HttpClient
-import io.ktor.client.HttpClientConfig
-import io.ktor.client.plugins.HttpRequestRetry
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.client.plugins.logging.LogLevel
-import io.ktor.client.plugins.logging.Logger
-import io.ktor.client.plugins.logging.Logging
-import io.ktor.serialization.kotlinx.json.json
 import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.JsonBuilder
 
 /**
  * Alias that denotes a forward geocoding [HttpApiEndpoint].
@@ -30,65 +22,7 @@ public typealias ReverseEndpoint = HttpApiEndpoint<Coordinates, List<Place>>
  */
 public interface HttpApiPlatformGeocoder : PlatformGeocoder {
 
-    public companion object {
-
-        /**
-         * Create a new [Json] configuration with the given [block].
-         *
-         * @param block Customization of the [Json] configuration.
-         * @return A new [Json] configuration with the given [block].
-         */
-        public fun json(block: JsonBuilder.() -> Unit = {}): Json = Json {
-            ignoreUnknownKeys = true
-            apply(block)
-        }
-
-        /**
-         * Create a new [HttpClient] with the given [json] configuration.
-         *
-         * Customization of the [HttpClient] can be done with the [block] parameter.
-         *
-         * @param json The [Json] configuration to use for the [HttpClient].
-         * @param enableLogging Whether or not to enable logging.
-         * @param enableRetry Whether or not to enable retrying 500 errors.
-         * @param maxRetries The maximum number of retries to attempt.
-         * @param block Customization of the [HttpClient].
-         * @return A new [HttpClient] with the given [json] configuration.
-         */
-        public fun httpClient(
-            json: Json = json(),
-            enableLogging: Boolean = false,
-            enableRetry: Boolean = true,
-            maxRetries: Int = 3,
-            block: HttpClientConfig<*>.() -> Unit = {},
-        ): HttpClient = HttpClient {
-            expectSuccess = true
-
-            if (enableRetry) {
-                install(HttpRequestRetry) {
-                    retryOnServerErrors(maxRetries = maxRetries)
-                    exponentialDelay()
-                }
-            }
-
-            if (enableLogging) {
-                install(Logging) {
-                    level = LogLevel.ALL
-                    logger = object : Logger {
-                        override fun log(message: String) {
-                            co.touchlab.kermit.Logger.i { message }
-                        }
-                    }
-                }
-            }
-
-            install(ContentNegotiation) {
-                json(json)
-            }
-
-            apply(block)
-        }
-    }
+    public companion object
 }
 
 /**
@@ -106,8 +40,8 @@ public interface HttpApiPlatformGeocoder : PlatformGeocoder {
 public fun HttpApiPlatformGeocoder(
     forwardEndpoint: ForwardEndpoint,
     reverseEndpoint: ReverseEndpoint,
-    json: Json = HttpApiPlatformGeocoder.json(),
-    client: HttpClient = HttpApiPlatformGeocoder.httpClient(json),
+    json: Json = HttpApiEndpoint.json(),
+    client: HttpClient = HttpApiEndpoint.httpClient(json),
 ): HttpApiPlatformGeocoder = object : HttpApiPlatformGeocoder {
 
     override fun isAvailable(): Boolean = true
