+++ compass-tools-web/src/commonMain/kotlin/dev/jordond/compass/tools/web/HttpApiEndpoint.kt
@@ -1,6 +1,16 @@
 package dev.jordond.compass.tools.web
 
+import io.ktor.client.HttpClient
+import io.ktor.client.HttpClientConfig
+import io.ktor.client.plugins.HttpRequestRetry
+import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.client.plugins.logging.LogLevel
+import io.ktor.client.plugins.logging.Logger
+import io.ktor.client.plugins.logging.Logging
 import io.ktor.client.statement.HttpResponse
+import io.ktor.serialization.kotlinx.json.json
+import kotlinx.serialization.json.Json
+import kotlinx.serialization.json.JsonBuilder
 
 /**
  * Defines an HTTP API endpoint.
@@ -25,6 +35,66 @@ public interface HttpApiEndpoint<Params, Result> {
      * @return The result type.
      */
     public suspend fun mapResponse(response: HttpResponse): Result
+
+    public companion object {
+
+        /**
+         * Create a new [Json] configuration with the given [block].
+         *
+         * @param block Customization of the [Json] configuration.
+         * @return A new [Json] configuration with the given [block].
+         */
+        public fun json(block: JsonBuilder.() -> Unit = {}): Json = Json {
+            ignoreUnknownKeys = true
+            apply(block)
+        }
+
+        /**
+         * Create a new [HttpClient] with the given [json] configuration.
+         *
+         * Customization of the [HttpClient] can be done with the [block] parameter.
+         *
+         * @param json The [Json] configuration to use for the [HttpClient].
+         * @param enableLogging Whether or not to enable logging.
+         * @param enableRetry Whether or not to enable retrying 500 errors.
+         * @param maxRetries The maximum number of retries to attempt.
+         * @param block Customization of the [HttpClient].
+         * @return A new [HttpClient] with the given [json] configuration.
+         */
+        public fun httpClient(
+            json: Json = json(),
+            enableLogging: Boolean = false,
+            enableRetry: Boolean = true,
+            maxRetries: Int = 3,
+            block: HttpClientConfig<*>.() -> Unit = {},
+        ): HttpClient = HttpClient {
+            expectSuccess = true
+
+            if (enableRetry) {
+                install(HttpRequestRetry) {
+                    retryOnServerErrors(maxRetries = maxRetries)
+                    exponentialDelay()
+                }
+            }
+
+            if (enableLogging) {
+                install(Logging) {
+                    level = LogLevel.ALL
+                    logger = object : Logger {
+                        override fun log(message: String) {
+                            co.touchlab.kermit.Logger.i { message }
+                        }
+                    }
+                }
+            }
+
+            install(ContentNegotiation) {
+                json(json)
+            }
+
+            apply(block)
+        }
+    }
 }
 
 /**
