+++ demo/composeApp/src/commonMain/kotlin/geocoder/GoogleMapsGeocoderScreen.kt
@@ -9,8 +9,8 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import cafe.adriel.voyager.core.screen.Screen
 import dev.jordond.compass.geocoder.web.GoogleMapsGeocoder
-import dev.jordond.compass.geocoder.web.HttpApiPlatformGeocoder
 import dev.jordond.compass.geocoder.web.parameter.GoogleMapsLocationType
+import dev.jordond.compass.tools.web.HttpApiEndpoint
 
 /**
  * This screen demonstrates how to use the Google Maps Geocoder for all platforms.
@@ -24,7 +24,7 @@ class GoogleMapsGeocoderScreen : Screen {
             derivedStateOf {
                 GoogleMapsGeocoder(
                     apiKey = apiKey,
-                    client = HttpApiPlatformGeocoder.httpClient(enableLogging = true)
+                    client = HttpApiEndpoint.httpClient(enableLogging = true)
                 ) {
                     locationType(GoogleMapsLocationType.Approximate)
                 }
